Move scene param types to their respective fiat plugin scene files
